backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
requests==2.31.0
python-dotenv==1.0.0
werkzeug==2.3.7
# pandas==2.1.0  # We'll add this later when needed